Added gitignore and organized classes
Added the gitignore to ignore __pycache and pyc files, and then put all of our classes in a folder to organize them
