+++ classes/Main.py
@@ -1,0 +1,86 @@
+import importlib
+import sys
+
+## Tkinter and Tkinter Framework
+import tkinter as tk
+from tkinter import Scrollbar, ttk
+import ttkbootstrap as tkb
+from ttkbootstrap.constants import *
+from ttkbootstrap.style import Bootstyle
+
+## Image imports
+from PIL import Image, ImageTk
+from urllib.request import urlopen
+import sqlite3
+
+class App(tk.Tk):
+    def __init__(self):
+        super().__init__()
+        self.title("Kroger")
+        self.state("zoomed")
+        self.conn = sqlite3.connect('assets/PyProject.db')
+        self.cursor = self.conn.cursor()
+        self.createNavbar()
+        self.current_page = None      
+        self.openPage("homePage")
+
+    def destroyPage(self):
+        if self.current_page:
+            self.current_page.destroy()
+
+    def load_page(self, page_name):
+        try:
+            # import the module using the full path
+            module = importlib.import_module(f"pages.{page_name}")
+            # get the page class from the module
+            page_class = getattr(module, page_name)
+            return page_class
+        except (ImportError, AttributeError):
+            print(f"Error: Page {page_name} does not exist.")
+            return None
+        
+    def openPage(self, page_name, *args):
+        page_class = self.load_page(page_name)
+        if page_class is None:
+            return
+
+        if not isinstance(self.current_page, page_class):
+            if self.current_page != None:
+                self.current_page.pack_forget()  # Use pack_forget instead of destroy
+            self.current_page = page_class(self, self, *args)
+            self.current_page.pack(fill=BOTH, expand=True)  # Add fill and expand options for better page display
+
+
+    def createNavbar(self):
+        # create the top menu
+        self.navbar = tkb.Frame(self, bootstyle="primary")
+        self.navbar.pack(fill=X, ipady=10, side=TOP)
+
+        #logo frame
+        self.logoFrame = tkb.Frame(self.navbar, bootstyle="primary")
+        self.logoFrame.pack(fill=Y, side=LEFT)
+        #logo
+        self.logo = tkb.Button(self.logoFrame, width=10, text="Kroger", command=lambda: self.openPage("homePage"))
+        self.logo.pack(side=LEFT, ipady=8, padx=2)
+
+        # Search frame
+        self.searchFrame = tkb.Frame(self.navbar, bootstyle="primary")
+        self.searchFrame.pack(fill=Y, side=LEFT)
+        # Search bar and button
+        self.searchBar = tkb.Entry(self.searchFrame, width=180, bootstyle="secondary")
+        self.searchBar.pack(side=LEFT, ipady=8, padx=2)
+        self.searchButton = tkb.Button(master=self.searchFrame, text='Search', compound=RIGHT)
+        self.searchButton.pack(fill=Y, side=RIGHT)
+
+        # user frame
+        self.userFrame = tkb.Frame(self.navbar, bootstyle="primary")
+        self.userFrame.pack(fill=Y, side=RIGHT)
+        # user / cart button
+        self.cartButton  = ttk.Button(master=self.userFrame, text='Cart', compound=RIGHT, command=lambda: self.openPage("cartPage"))
+        self.cartButton.pack(side=LEFT, ipadx=10, fill=Y)
+        self.userButton  = ttk.Button(master=self.userFrame, text='User', compound=RIGHT, command=lambda: self.openPage("userPage"))
+        self.userButton.pack(side=LEFT, ipadx=10, fill=Y) 
+
+if __name__ == "__main__":
+    app = App()
+    app.mainloop()
